backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk>=6.0.1",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "gtts>=2.5.4",
    "isort>=8.0.1",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pylint>=4.0.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_decorators = false

[tool.pylint.messages_control]
max-line-length = 100
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=. --cov-report=term-missing -m 'not live'"
markers = [
    "live: tests that call external APIs (run with: pytest -m live)",
]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", ".venv/*"]
